Добавил обработку ошибок в контроллере API. Также, создал методы получения случайного пользователи и пользователя по Id на клиенте

+++ Lecture_No7_HW/Application/UseCases/RemoveCustomerUseCase.cs
@@ -1,4 +1,5 @@
-﻿using Application.UseCases.Contracts;
+﻿using Application.Errors;
+using Application.UseCases.Contracts;
 using Domain.Common.Repositories.Abstractions;
 
 namespace Application.UseCases
@@ -14,6 +15,9 @@ namespace Application.UseCases
 
         public async Task EcxecuteAsync(int id)
         {
+            var foundCustomer = await customerRepository.GetAsync(id);
+            if (foundCustomer != null) 
+                throw new CustomerNotFoundExeption($"Пользователь с идентификатором №{id} не найден");
             await customerRepository.DeleteAsync(id);
         }
     }
